requirements.txt
Flask==3.1.0
beautifulsoup4==4.12.3
requests==2.32.3
selenium==4.32.0
SQLAlchemy==2.0.40
python-dotenv==1.1.0
psycopg2-binary==2.9.10
pytz==2024.2
gunicorn